+++ kite/src/main/kotlin/com/tang/kite/sql/TableReference.kt
@@ -1,5 +1,6 @@
 package com.tang.kite.sql
 
+import com.tang.kite.constants.SqlString
 import com.tang.kite.utils.Reflects
 import kotlin.reflect.KClass
 
@@ -10,16 +11,24 @@ class TableReference(
 
     val name: String,
 
-    var clazz: Class<*>? = null,
+    val clazz: Class<*>,
 
-    val alias: String? = null,
+    val alias: String,
 
     val schema: String? = null
 
 ) {
 
-    constructor(clazz: Class<*>) : this(name = Reflects.getTableName(clazz), clazz = clazz)
+    constructor(clazz: Class<*>) : this(Reflects.getTableName(clazz), clazz, Reflects.getTableAlias(clazz))
 
     constructor(clazz: KClass<*>) : this(clazz.java)
+
+    fun toString(withAlias: Boolean): String {
+        return if (withAlias) {
+            name + SqlString.AS + alias
+        } else {
+            name
+        }
+    }
 
 }