+++ kite/src/test/kotlin/com/tang/kite/sql/SqlNodeTest.kt
@@ -1,8 +1,10 @@
 package com.tang.kite.sql
 
 import com.tang.kite.session.entity.Account
+import com.tang.kite.sql.dialect.SqlDialect
 import com.tang.kite.sql.enumeration.ComparisonOperator
 import com.tang.kite.sql.enumeration.LogicalOperator
+import com.tang.kite.sql.provider.ProviderType
 import com.tang.kite.sql.statement.ComparisonStatement
 import com.tang.kite.sql.statement.LogicalStatement
 import kotlin.test.Test
@@ -22,8 +24,14 @@ class SqlNodeTest {
         val condition = ComparisonStatement(Column(Account::id), 1, ComparisonOperator.EQUAL)
         select.where.add(LogicalStatement(condition, LogicalOperator.AND))
 
+        println(select.getSqlStatement(object : SqlDialect {
+            override fun getType(): ProviderType {
+                return ProviderType.POSTGRESQL
+            }
 
-        println()
+            override fun applyLimitClause(sql: StringBuilder, parameters: MutableList<Any?>, limit: Long, offset: Long) {
+            }
+        }))
     }
 
 }